Updated incoming messages

+++ Terraria_Server/Collections/IRegisterableEntity.cs
@@ -8,7 +8,7 @@ namespace Terraria_Server.Collections
     public interface IRegisterableEntity : ICloneable
     {
         int Type { get; set; }
-        short NetID { get; set; }
+        int NetID { get; set; }
         string Name { get; set; }
         bool Active { get; set; }
     }
